Fixed bug when bound element had child elements.
When mousestop was bound to an element that contained child elements, the mouseenter and mouseleave events
were not always fired. If the child elements took up the entire portion of the parent element, the events were
never fired.

I replaced mouseenter and mouseleave with mouseover and mouseout respectively to alleviate this problem.

Additionally, using $(e.target) would return the incorrect element since it may point to the child element.
Instead, we now use $(this) to ensure we always have the element to which mousestop was bound.